SLCORE-2062 Move binaries utilities to plugin/source package
OnDemandPluginCacheManager, OmnisharpDistributionDownloader,
OnDemandPluginSignatureVerifier, UniqueTaskExecutor and DownloadableArtifact
were pure package moves; delete the old files and update imports in
SonarLintSpringAppConfig and AnalysisSchedulerCache.
