+++ backend/core/src/main/java/org/sonarsource/sonarlint/core/spring/SonarLintSpringAppConfig.java
@@ -96,14 +96,14 @@ import org.sonarsource.sonarlint.core.plugin.ServerPluginsCache;
 import org.sonarsource.sonarlint.core.plugin.resolvers.ConnectedModeArtifactResolver;
 import org.sonarsource.sonarlint.core.plugin.resolvers.ConnectedModeCompanionPluginResolver;
 import org.sonarsource.sonarlint.core.plugin.resolvers.EmbeddedArtifactResolver;
-import org.sonarsource.sonarlint.core.plugin.resolvers.OmnisharpDistributionDownloader;
 import org.sonarsource.sonarlint.core.plugin.resolvers.OnDemandArtifactResolver;
-import org.sonarsource.sonarlint.core.plugin.resolvers.OnDemandPluginCacheManager;
-import org.sonarsource.sonarlint.core.plugin.resolvers.OnDemandPluginSignatureVerifier;
 import org.sonarsource.sonarlint.core.plugin.resolvers.PluginOverrideRegistry;
 import org.sonarsource.sonarlint.core.plugin.resolvers.PremiumArtifactResolver;
 import org.sonarsource.sonarlint.core.plugin.resolvers.ServerPluginDownloader;
-import org.sonarsource.sonarlint.core.plugin.resolvers.UniqueTaskExecutor;
+import org.sonarsource.sonarlint.core.plugin.source.UniqueTaskExecutor;
+import org.sonarsource.sonarlint.core.plugin.source.binaries.OmnisharpDistributionDownloader;
+import org.sonarsource.sonarlint.core.plugin.source.binaries.OnDemandPluginCacheManager;
+import org.sonarsource.sonarlint.core.plugin.source.binaries.OnDemandPluginSignatureVerifier;
 import org.sonarsource.sonarlint.core.plugin.resolvers.UnsupportedArtifactResolver;
 import org.sonarsource.sonarlint.core.plugin.skipped.SkippedPluginsNotifierService;
 import org.sonarsource.sonarlint.core.plugin.skipped.SkippedPluginsRepository;
