+++ backend/core/src/test/java/org/sonarsource/sonarlint/core/analysis/AnalysisSchedulerCacheTest.java
@@ -26,7 +26,7 @@ import org.sonarsource.sonarlint.core.UserPaths;
 import org.sonarsource.sonarlint.core.fs.ClientFileSystemService;
 import org.sonarsource.sonarlint.core.plugin.PluginLifecycleService;
 import org.sonarsource.sonarlint.core.plugin.PluginsService;
-import org.sonarsource.sonarlint.core.plugin.resolvers.OmnisharpDistributionDownloader;
+import org.sonarsource.sonarlint.core.plugin.source.binaries.OmnisharpDistributionDownloader;
 import org.sonarsource.sonarlint.core.repository.config.ConfigurationRepository;
 
 import static org.mockito.Mockito.mock;
